Making progress

+++ class-generator/src/main/java/com/sandflow/smpte/mxf/MXFOutputContext.java
@@ -7,6 +7,6 @@ import com.sandflow.smpte.util.UUID;
 
 public interface MXFOutputContext {
   UUID getPackageInstanceID(UMID packageID);
-  long getLocalTag(AUID auid);
+  int getLocalTag(AUID auid);
   void putSet(Set set);
 }
